+++ lib/processor/process_dag.go
@@ -55,8 +55,7 @@ process_dag:
       .: .
     processors:
     - http:
-        request:
-          url: http://foo/enrich
+        url: http://foo/enrich
     postmap:
       foo_result: .
 
@@ -65,8 +64,7 @@ process_dag:
       .: msg.sub.path
     processors:
     - http:
-        request:
-          url: http://bar/enrich
+        url: http://bar/enrich
     postmap:
       bar_result: .
 
@@ -76,8 +74,7 @@ process_dag:
       bar_obj: bar_result
     processors:
     - http:
-        request:
-          url: http://baz/enrich
+        url: http://baz/enrich
     postmap:
       baz_obj: .
 ` + "```" + `